+++ js/script.js
@@ -1,0 +1,10 @@
+
+/*Bloquear scroll si no se le hace click dentro del mapa*/
+
+ $('.googlemap').click(function () {
+    $('.googlemap iframe').css("pointer-events", "auto");
+});
+
+$( ".googlemap" ).mouseleave(function() {
+  $('.googlemap iframe').css("pointer-events", "none"); 
+});