+++ CHANGELOG.md
@@ -1,4 +1,34 @@
-## [1.0.0] - 2017-06-20
+## [1.0.0] - 2020-09-14
 ### Added 
 - added new section name changelog
 - added new section name .gitignore
+
+## [1.0.0] - 2020-09-17
+### Added
+-added files for the project (bootstrap, js, css, html)
+
+## [1.0.0] - 2020-09-20
+### Added
+-Oscar added new section Products
+
+## [1.0.0] - 2020-09-21
+### Changed
+-corrected some mistakes on the Products section
+
+## [1.0.0] - 2020-09-27
+### Added
+-added new section name contact
+
+## [1.0.0] - 2020-10-03
+### Changed
+-changed the products section
+
+## [1.0.0] - 2020-10-16
+### Added
+-added js to block map until you click on it
+-added css to block map on scrolling
+
+## [1.0.0] - 2020-10-16
+### Changed
+-changed index.html to index.php
+-changes form.php